ability to load 2Mb+ files, various test scenarious

+++ SPMeta2/SPMeta2.Regression/Services/RandomService.cs
@@ -11,12 +11,12 @@ namespace SPMeta2.Regression.Services
         public abstract Guid Guid();
 
         public abstract byte[] Content();
-        public abstract byte[] Content(int lenght);
+        public abstract byte[] Content(long lenght);
 
         public abstract bool Bool();
 
         public abstract string String();
-        public abstract string String(int lenght);
+        public abstract string String(long lenght);
 
         public abstract int Int();
         public abstract int Int(int maxValue);
